Concise formatting when no improvements
Ensured that concise formatter behaves gracefully when there is no improvements to show in release notes.
This should be an extremely rare case - the point of notable releases is includes multiple improvements, making the release _notable_. However, I wanted to make sure the release notes content look good in this case. Previously, it showed empty space in release notes and it might have looked confusing / like bug to our users.