Removed travisRelease task
1. This task was hard to maintain / work with.
Forking off Gradle builds from inside Gradle build is problematic
because it is hard to fine tune the child Gradle build.
For example, there is no easy way to run the child Gradle build
with some extra flags like '-s' or '-i'.

2. I removed the task completely. Instead of this CI configuration:

./gradlew travisRelease

The CI build is configured this way:

./gradlew assertReleaseNeeded && ./gradlew travisReleasePrepare && ./gradlew performRelease

This is more code, but it is way more flexible to work with,
makes it easy to reproduce builds by running individual tasks.
Also, it is very explicit, removes the "magic" from the process.

3. New "assertReleaseNeeded" task fails when release is not needed.
This way, we can stop executing further steps.
This is useful when configuring CI builds.