Fixed problems with git push
Git push was broken after my previous refactorings.
It was hard to diagnose the problem because logging and error messages
from git push were hidden from the client build.
We want to hide sensitive information such as GitHub write auth token.
However, we want to maintain high level of debuggability.
This change does both.

git push output and the command are now printed to the console
but the sensitive information is hidden from the user.