Removed 'testRelease' task, release dry by default
1. testRelease task was problematic because it was forking Gradle build from Gradle build.
This approach turned out to be impractical because the logging from builds gets mixed up.
Also, there is no way to control the child build and pass any parameters like '-i' logging level.
2. testing release now is just running './gradlew performRelease' or './gradlew performRelease releaseCleanUp'
3. changed the default of release to be dryRun=true.
This way it is safer (no accidental releases).
Also, it is easier to test without forgetting to pass project property.