Removed test that I think is not necessary
I thought that having this test is nice but unfortunately we cannot static state like that.
Static state is preserved in Gradle daemon and it's better not to mess with the daemon :)

Given that we cannot keep static state testing this code is pretty cumbersome to test.
Since there is not much complexity in the code under test perhaps it's best to remove it.
If this code is broken, our integ tests will start failing badly, so I doubt we will have any bugs here.
We don't have yet integration tests, but soon :)