Started using mustRunAfter instead of shouldRunAfter
shouldRunAfter is technically more correct here because the ordering rules between bintray upload and git push is not a hard requirement. However, shouldRunAfter does not work correctly with lower versions of Gradle that we still support. It does not make the tasks have correct order. Hence, I'm using "mustRunAfter" for task ordering rules.