Lenient way of access shipkit configuration
By default, release configuration object fails fast if one tries to access a property that was not set before (e.g. is null). This is a good default because it helps us fleshing configuration issues early. However sometimes we want to check if the user has configured a property (like for GitHub write token) without failing. The new "getLenient()" public method offers a copy of release configuration that does not fail fast when requested property is null.