Improved the default prefix of forked processes
I found it not easy to debug if the default prefix was the same for every invocation.
Now, instead of [./gradlew] we will see [performRelease], instead of [git] (git status) we will see [status].
I think it is a better default, it will be easier to debug Travis CI builds.